requirements.txt
streamlit>=1.28.0
pypdf>=3.17.0
requests>=2.31.0
python-dotenv>=1.0.0
